requirements.txt
attrs>21
pymodbus
# pymodbus[repl]
paho-mqtt~=1.5.0
pyserial-asyncio
